test: Update makeMockChainStorageRoot for proper StreamCell handling
