+++ packages/vats/test/test-board-utils.js
@@ -11,11 +11,14 @@ import {
   makeHistoryReviver,
 } from '../tools/board-utils.js';
 
+const streamCellTextFromArray = values =>
+  JSON.stringify({ blockHeight: '0', values });
+
 /** @type {Array<[string, any]>} */
 const agoricNamesDataEntriesFixture = [
   [
     'published.agoricNames.issuer',
-    [
+    streamCellTextFromArray([
       '{"body":"[]","slots":[]}',
       '{"body":"[[\\"IST\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST issuer\\",\\"index\\":0}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation issuer\\",\\"index\\":1}]]","slots":["board0371","board0592"]}',
       '{"body":"[[\\"BLD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD issuer\\",\\"index\\":0}],[\\"IST\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST issuer\\",\\"index\\":1}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation issuer\\",\\"index\\":2}]]","slots":["board0425","board0371","board0592"]}',
@@ -25,11 +28,11 @@ const agoricNamesDataEntriesFixture = [
       '{"body":"[[\\"BLD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD issuer\\",\\"index\\":0}],[\\"IST\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST issuer\\",\\"index\\":1}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation issuer\\",\\"index\\":2}],[\\"USDC_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl issuer\\",\\"index\\":3}],[\\"USDC_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv issuer\\",\\"index\\":4}],[\\"USDT_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl issuer\\",\\"index\\":5}],[\\"USDT_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv issuer\\",\\"index\\":6}]]","slots":["board0425","board0371","board0592","board02314","board00917","board06120","board05024"]}',
       '{"body":"[[\\"BLD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD issuer\\",\\"index\\":0}],[\\"IST\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST issuer\\",\\"index\\":1}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation issuer\\",\\"index\\":2}],[\\"AUSD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: AUSD issuer\\",\\"index\\":3}],[\\"USDC_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl issuer\\",\\"index\\":4}],[\\"USDC_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv issuer\\",\\"index\\":5}],[\\"USDT_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl issuer\\",\\"index\\":6}],[\\"USDT_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv issuer\\",\\"index\\":7}]]","slots":["board0425","board0371","board0592","board04827","board02314","board00917","board06120","board05024"]}',
       '{"body":"[[\\"BLD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD issuer\\",\\"index\\":0}],[\\"IST\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST issuer\\",\\"index\\":1}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation issuer\\",\\"index\\":2}],[\\"AUSD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: AUSD issuer\\",\\"index\\":3}],[\\"ATOM\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: ATOM issuer\\",\\"index\\":4}],[\\"USDC_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl issuer\\",\\"index\\":5}],[\\"USDC_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv issuer\\",\\"index\\":6}],[\\"USDT_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl issuer\\",\\"index\\":7}],[\\"USDT_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv issuer\\",\\"index\\":8}]]","slots":["board0425","board0371","board0592","board04827","board02575","board02314","board00917","board06120","board05024"]}',
-    ],
+    ]),
   ],
   [
     'published.agoricNames.brand',
-    [
+    streamCellTextFromArray([
       '{"body":"[]","slots":[]}',
       '{"body":"[[\\"BLD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation brand\\",\\"index\\":1}]]","slots":["board0223","board0074"]}',
       '{"body":"[[\\"BLD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0}],[\\"IST\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST brand\\",\\"index\\":1}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation brand\\",\\"index\\":2}]]","slots":["board0223","board0566","board0074"]}',
@@ -39,11 +42,11 @@ const agoricNamesDataEntriesFixture = [
       '{"body":"[[\\"BLD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0}],[\\"IST\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST brand\\",\\"index\\":1}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation brand\\",\\"index\\":2}],[\\"USDC_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl brand\\",\\"index\\":3}],[\\"USDC_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv brand\\",\\"index\\":4}],[\\"USDT_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl brand\\",\\"index\\":5}],[\\"USDT_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv brand\\",\\"index\\":6}]]","slots":["board0223","board0566","board0074","board05815","board00218","board02021","board03125"]}',
       '{"body":"[[\\"BLD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0}],[\\"IST\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST brand\\",\\"index\\":1}],[\\"AUSD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: AUSD brand\\",\\"index\\":2}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation brand\\",\\"index\\":3}],[\\"USDC_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl brand\\",\\"index\\":4}],[\\"USDC_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv brand\\",\\"index\\":5}],[\\"USDT_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl brand\\",\\"index\\":6}],[\\"USDT_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv brand\\",\\"index\\":7}]]","slots":["board0223","board0566","board03928","board0074","board05815","board00218","board02021","board03125"]}',
       '{"body":"[[\\"BLD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0}],[\\"IST\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST brand\\",\\"index\\":1}],[\\"AUSD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: AUSD brand\\",\\"index\\":2}],[\\"Invitation\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: Zoe Invitation brand\\",\\"index\\":3}],[\\"ATOM\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: ATOM brand\\",\\"index\\":4}],[\\"USDC_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl brand\\",\\"index\\":5}],[\\"USDC_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv brand\\",\\"index\\":6}],[\\"USDT_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl brand\\",\\"index\\":7}],[\\"USDT_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv brand\\",\\"index\\":8}]]","slots":["board0223","board0566","board03928","board0074","board02963","board05815","board00218","board02021","board03125"]}',
-    ],
+    ]),
   ],
   [
     'published.agoricNames.vbankAsset',
-    [
+    streamCellTextFromArray([
       '{"body":"[]","slots":[]}',
       '{"body":"[[\\"ubld\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0},\\"denom\\":\\"ubld\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD issuer\\",\\"index\\":1},\\"issuerName\\":\\"BLD\\",\\"proposedName\\":\\"Agoric staking token\\"}]]","slots":["board0223","board0425"]}',
       '{"body":"[[\\"ubld\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0},\\"denom\\":\\"ubld\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD issuer\\",\\"index\\":1},\\"issuerName\\":\\"BLD\\",\\"proposedName\\":\\"Agoric staking token\\"}],[\\"uist\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST brand\\",\\"index\\":2},\\"denom\\":\\"uist\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST issuer\\",\\"index\\":3},\\"issuerName\\":\\"IST\\",\\"proposedName\\":\\"Agoric stable token\\"}]]","slots":["board0223","board0425","board0566","board0371"]}',
@@ -53,11 +56,11 @@ const agoricNamesDataEntriesFixture = [
       '{"body":"[[\\"ubld\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0},\\"denom\\":\\"ubld\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD issuer\\",\\"index\\":1},\\"issuerName\\":\\"BLD\\",\\"proposedName\\":\\"Agoric staking token\\"}],[\\"uist\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST brand\\",\\"index\\":2},\\"denom\\":\\"uist\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST issuer\\",\\"index\\":3},\\"issuerName\\":\\"IST\\",\\"proposedName\\":\\"Agoric stable token\\"}],[\\"ibc/toyusdc\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl brand\\",\\"index\\":4},\\"denom\\":\\"ibc/toyusdc\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl issuer\\",\\"index\\":5},\\"issuerName\\":\\"USDC_axl\\",\\"proposedName\\":\\"USD Coin\\"}],[\\"ibc/usdc5678\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv brand\\",\\"index\\":6},\\"denom\\":\\"ibc/usdc5678\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv issuer\\",\\"index\\":7},\\"issuerName\\":\\"USDC_grv\\",\\"proposedName\\":\\"USC Coin\\"}],[\\"ibc/usdt1234\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl brand\\",\\"index\\":8},\\"denom\\":\\"ibc/usdt1234\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl issuer\\",\\"index\\":9},\\"issuerName\\":\\"USDT_axl\\",\\"proposedName\\":\\"Tether USD\\"}],[\\"ibc/toyollie\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv brand\\",\\"index\\":10},\\"denom\\":\\"ibc/toyollie\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv issuer\\",\\"index\\":11},\\"issuerName\\":\\"USDT_grv\\",\\"proposedName\\":\\"Tether USD\\"}]]","slots":["board0223","board0425","board0566","board0371","board05815","board02314","board00218","board00917","board02021","board06120","board03125","board05024"]}',
       '{"body":"[[\\"ubld\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0},\\"denom\\":\\"ubld\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD issuer\\",\\"index\\":1},\\"issuerName\\":\\"BLD\\",\\"proposedName\\":\\"Agoric staking token\\"}],[\\"uist\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST brand\\",\\"index\\":2},\\"denom\\":\\"uist\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST issuer\\",\\"index\\":3},\\"issuerName\\":\\"IST\\",\\"proposedName\\":\\"Agoric stable token\\"}],[\\"ibc/toyusdc\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl brand\\",\\"index\\":4},\\"denom\\":\\"ibc/toyusdc\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl issuer\\",\\"index\\":5},\\"issuerName\\":\\"USDC_axl\\",\\"proposedName\\":\\"USD Coin\\"}],[\\"ibc/usdc5678\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv brand\\",\\"index\\":6},\\"denom\\":\\"ibc/usdc5678\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv issuer\\",\\"index\\":7},\\"issuerName\\":\\"USDC_grv\\",\\"proposedName\\":\\"USC Coin\\"}],[\\"ibc/usdt1234\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl brand\\",\\"index\\":8},\\"denom\\":\\"ibc/usdt1234\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl issuer\\",\\"index\\":9},\\"issuerName\\":\\"USDT_axl\\",\\"proposedName\\":\\"Tether USD\\"}],[\\"ibc/toyollie\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv brand\\",\\"index\\":10},\\"denom\\":\\"ibc/toyollie\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv issuer\\",\\"index\\":11},\\"issuerName\\":\\"USDT_grv\\",\\"proposedName\\":\\"Tether USD\\"}],[\\"ibc/toyellie\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: AUSD brand\\",\\"index\\":12},\\"denom\\":\\"ibc/toyellie\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: AUSD issuer\\",\\"index\\":13},\\"issuerName\\":\\"AUSD\\",\\"proposedName\\":\\"Anchor USD\\"}]]","slots":["board0223","board0425","board0566","board0371","board05815","board02314","board00218","board00917","board02021","board06120","board03125","board05024","board03928","board04827"]}',
       '{"body":"[[\\"ubld\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD brand\\",\\"index\\":0},\\"denom\\":\\"ubld\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BLD issuer\\",\\"index\\":1},\\"issuerName\\":\\"BLD\\",\\"proposedName\\":\\"Agoric staking token\\"}],[\\"uist\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST brand\\",\\"index\\":2},\\"denom\\":\\"uist\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: IST issuer\\",\\"index\\":3},\\"issuerName\\":\\"IST\\",\\"proposedName\\":\\"Agoric stable token\\"}],[\\"ibc/toyusdc\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl brand\\",\\"index\\":4},\\"denom\\":\\"ibc/toyusdc\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_axl issuer\\",\\"index\\":5},\\"issuerName\\":\\"USDC_axl\\",\\"proposedName\\":\\"USD Coin\\"}],[\\"ibc/usdc5678\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv brand\\",\\"index\\":6},\\"denom\\":\\"ibc/usdc5678\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDC_grv issuer\\",\\"index\\":7},\\"issuerName\\":\\"USDC_grv\\",\\"proposedName\\":\\"USC Coin\\"}],[\\"ibc/usdt1234\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl brand\\",\\"index\\":8},\\"denom\\":\\"ibc/usdt1234\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_axl issuer\\",\\"index\\":9},\\"issuerName\\":\\"USDT_axl\\",\\"proposedName\\":\\"Tether USD\\"}],[\\"ibc/toyollie\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv brand\\",\\"index\\":10},\\"denom\\":\\"ibc/toyollie\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USDT_grv issuer\\",\\"index\\":11},\\"issuerName\\":\\"USDT_grv\\",\\"proposedName\\":\\"Tether USD\\"}],[\\"ibc/toyellie\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: AUSD brand\\",\\"index\\":12},\\"denom\\":\\"ibc/toyellie\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: AUSD issuer\\",\\"index\\":13},\\"issuerName\\":\\"AUSD\\",\\"proposedName\\":\\"Anchor USD\\"}],[\\"ibc/toyatom\\",{\\"brand\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: ATOM brand\\",\\"index\\":14},\\"denom\\":\\"ibc/toyatom\\",\\"displayInfo\\":{\\"assetKind\\":\\"nat\\",\\"decimalPlaces\\":6},\\"issuer\\":{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: ATOM issuer\\",\\"index\\":15},\\"issuerName\\":\\"ATOM\\",\\"proposedName\\":\\"ATOM\\"}]]","slots":["board0223","board0425","board0566","board0371","board05815","board02314","board00218","board00917","board02021","board06120","board03125","board05024","board03928","board04827","board02963","board02575"]}',
-    ],
+    ]),
   ],
   [
     'published.agoricNames.installation',
-    [
+    streamCellTextFromArray([
       '{"body":"[]","slots":[]}',
       '{"body":"[[\\"centralSupply\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":0}],[\\"mintHolder\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":1}],[\\"provisionPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":2}]]","slots":["board0257","board0188","board0639"]}',
       '{"body":"[[\\"centralSupply\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":0}],[\\"mintHolder\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":1}],[\\"walletFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":2}],[\\"provisionPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":3}]]","slots":["board0257","board0188","board04312","board0639"]}',
@@ -103,11 +106,11 @@ const agoricNamesDataEntriesFixture = [
       '{"body":"[[\\"centralSupply\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":0}],[\\"mintHolder\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":1}],[\\"walletFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":2}],[\\"provisionPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":3}],[\\"contractGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":4}],[\\"committee\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":5}],[\\"binaryVoteCounter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":6}],[\\"amm\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":7}],[\\"VaultFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":8}],[\\"feeDistributor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":9}],[\\"liquidate\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":10}],[\\"stakeFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":11}],[\\"reserve\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":12}],[\\"psm\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":13}],[\\"econCommitteeCharter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":14}],[\\"interchainPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":15}],[\\"priceAggregator\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":16}],[\\"scaledPriceAuthority\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":17}]]","slots":["board0257","board05674","board04312","board0639","board04154","board02656","board06458","board01029","board00530","board04431","board02733","board02437","board03935","board01272","board01151","board05736","board00776","board01759"]}',
       '{"body":"[[\\"centralSupply\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":0}],[\\"mintHolder\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":1}],[\\"walletFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":2}],[\\"provisionPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":3}],[\\"contractGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":4}],[\\"committee\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":5}],[\\"binaryVoteCounter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":6}],[\\"amm\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":7}],[\\"VaultFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":8}],[\\"feeDistributor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":9}],[\\"liquidate\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":10}],[\\"stakeFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":11}],[\\"reserve\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":12}],[\\"psm\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":13}],[\\"econCommitteeCharter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":14}],[\\"interchainPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":15}],[\\"priceAggregator\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":16}],[\\"scaledPriceAuthority\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":17}]]","slots":["board0257","board05674","board04312","board0639","board04154","board02656","board06458","board01029","board00530","board04431","board02733","board02437","board03935","board01272","board01151","board05736","board00776","board01759"]}',
       '{"body":"[[\\"centralSupply\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":0}],[\\"mintHolder\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":1}],[\\"walletFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":2}],[\\"provisionPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":3}],[\\"contractGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":4}],[\\"committee\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":5}],[\\"binaryVoteCounter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":6}],[\\"amm\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":7}],[\\"VaultFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":8}],[\\"feeDistributor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":9}],[\\"liquidate\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":10}],[\\"stakeFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":11}],[\\"reserve\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":12}],[\\"psm\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":13}],[\\"econCommitteeCharter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":14}],[\\"interchainPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":15}],[\\"priceAggregator\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":16}],[\\"scaledPriceAuthority\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: BundleIDInstallation\\",\\"index\\":17}]]","slots":["board0257","board05674","board04312","board0639","board04154","board02656","board06458","board01029","board00530","board04431","board02733","board02437","board03935","board01272","board04277","board05736","board00776","board01759"]}',
-    ],
+    ]),
   ],
   [
     'published.agoricNames.instance',
-    [
+    streamCellTextFromArray([
       '{"body":"[]","slots":[]}',
       '{"body":"[[\\"econCommitteeCharter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":0}]]","slots":["board00613"]}',
       '{"body":"[[\\"economicCommittee\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":0}],[\\"econCommitteeCharter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":1}]]","slots":["board04016","board00613"]}',
@@ -127,16 +130,19 @@ const agoricNamesDataEntriesFixture = [
       '{"body":"[[\\"economicCommittee\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":0}],[\\"amm\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":1}],[\\"ammGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":2}],[\\"VaultFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":3}],[\\"reserve\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":4}],[\\"reserveGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":5}],[\\"econCommitteeCharter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":6}],[\\"interchainPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":7}],[\\"provisionPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":8}],[\\"walletFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":9}],[\\"ATOM-USD price feed\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":10}],[\\"psm-IST-USDC_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":11}],[\\"psm-IST-USDC_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":12}],[\\"psm-IST-USDT_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":13}],[\\"psm-IST-USDT_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":14}],[\\"psm-IST-AUSD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":15}]]","slots":["board04016","board00855","board00443","board05970","board02152","board06053","board00613","board05557","board01744","board03523","board06445","board03446","board01547","board00848","board04149","board03850"]}',
       '{"body":"[[\\"economicCommittee\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":0}],[\\"amm\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":1}],[\\"ammGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":2}],[\\"VaultFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":3}],[\\"Treasury\\",{\\"@qclass\\":\\"slot\\",\\"index\\":3}],[\\"reserve\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":4}],[\\"reserveGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":5}],[\\"econCommitteeCharter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":6}],[\\"interchainPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":7}],[\\"provisionPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":8}],[\\"walletFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":9}],[\\"ATOM-USD price feed\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":10}],[\\"psm-IST-USDC_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":11}],[\\"psm-IST-USDC_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":12}],[\\"psm-IST-USDT_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":13}],[\\"psm-IST-USDT_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":14}],[\\"psm-IST-AUSD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":15}]]","slots":["board04016","board00855","board00443","board05970","board02152","board06053","board00613","board05557","board01744","board03523","board06445","board03446","board01547","board00848","board04149","board03850"]}',
       '{"body":"[[\\"economicCommittee\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":0}],[\\"amm\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":1}],[\\"ammGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":2}],[\\"VaultFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":3}],[\\"Treasury\\",{\\"@qclass\\":\\"slot\\",\\"index\\":3}],[\\"VaultFactoryGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":4}],[\\"reserve\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":5}],[\\"reserveGovernor\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":6}],[\\"econCommitteeCharter\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":7}],[\\"interchainPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":8}],[\\"provisionPool\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":9}],[\\"walletFactory\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":10}],[\\"ATOM-USD price feed\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":11}],[\\"psm-IST-USDC_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":12}],[\\"psm-IST-USDC_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":13}],[\\"psm-IST-USDT_axl\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":14}],[\\"psm-IST-USDT_grv\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":15}],[\\"psm-IST-AUSD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: InstanceHandle\\",\\"index\\":16}]]","slots":["board04016","board00855","board00443","board05970","board03773","board02152","board06053","board00613","board05557","board01744","board03523","board06445","board03446","board01547","board00848","board04149","board03850"]}',
-    ],
+    ]),
   ],
   [
     'published.agoricNames.oracleBrand',
-    [
+    streamCellTextFromArray([
       '{"body":"[]","slots":[]}',
       '{"body":"[[\\"USD\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: USD brand\\",\\"index\\":0}],[\\"ATOM\\",{\\"@qclass\\":\\"slot\\",\\"iface\\":\\"Alleged: ATOM brand\\",\\"index\\":1}]]","slots":["board02810","board05311"]}',
-    ],
+    ]),
   ],
-  ['published.agoricNames.uiConfig', ['{"body":"[]","slots":[]}']],
+  [
+    'published.agoricNames.uiConfig',
+    streamCellTextFromArray(['{"body":"[]","slots":[]}']),
+  ],
 ];
 
 test('makeAgoricNamesRemotesFromFakeStorage', t => {
